+++ client/src/containers/Profile/actions.ts
@@ -1,10 +1,8 @@
 import * as authService from 'services/authService';
+import * as forgotPasswordService from 'services/forgotPasswordService';
 import { User } from 'types/user.type';
-import {
-  LoginCredentials,
-  RegisterCredentials,
-  ForgotPasswordCredentials,
-} from 'types/auth.types';
+import { LoginCredentials, RegisterCredentials } from 'types/auth.types';
+import { ForgotPasswordCredentials } from 'types/forgot.password.types';
 
 import { SET_USER, SET_IS_LOADING, AsyncUserAction, UserAction } from './action.type';
 
@@ -43,8 +41,10 @@ export const login = (request: LoginCredentials) =>
 export const registration = (request: RegisterCredentials) =>
   handleAuthResponse(authService.registration(request));
 
-export const forgotPassword = (request: ForgotPasswordCredentials) =>
-  handleAuthResponse(authService.forgotPassword(request));
+export const forgotPassword = (request: ForgotPasswordCredentials) => async () => {
+  const result = await forgotPasswordService.forgotPassword(request);
+  return result;
+};
 
 export const logout = () => setAuthData();
