+++ client/src/types/forgot.password.types.ts
@@ -1,0 +1,3 @@
+export type ForgotPasswordCredentials = {
+  email: string;
+};
